requirements.txt
#put all python dependencies in here.
picamera < 1.13
paho-mqtt
adafruit_python_dht
RPi.GPIO